refactor(hydro_lang): allow `sim` feature without `deploy` feature
Also removes leftover prototype "properties" code.
