+++ .github/workflows/ci.yml
@@ -110,13 +110,13 @@ jobs:
         env:
           INSTA_UPDATE: ${{ (matrix.rust_release == 'latest-nightly' && (github.event_name == 'schedule' || github.event_name == 'workflow_dispatch')) && 'always' || '' }}
           INSTA_FORCE_PASS: ${{ (matrix.rust_release == 'latest-nightly' && (github.event_name == 'schedule' || github.event_name == 'workflow_dispatch')) && '1' || '' }}
-        run: cargo nextest run --no-fail-fast --features hydro_lang/deploy --all-targets
+        run: cargo nextest run --no-fail-fast --all-targets
 
       - name: Run doctests
         id: doctests
         continue-on-error: ${{ matrix.rust_release == 'latest-nightly' && (github.event_name == 'schedule' || github.event_name == 'workflow_dispatch') }}
         # Not supported by nextest: https://github.com/nextest-rs/nextest/issues/16
-        run: cargo test --no-fail-fast --features hydro_lang/deploy --doc
+        run: cargo test --no-fail-fast --doc
 
       # Nightly snapshot update logic
       - name: Check for snapshot changes (nightly only)
