+++ hydro_lang/src/compile/mod.rs
@@ -6,11 +6,6 @@ pub mod ir;
 #[cfg(feature = "build")]
 #[cfg_attr(docsrs, doc(cfg(feature = "build")))]
 #[expect(missing_docs, reason = "TODO")]
-pub mod rewrites;
-
-#[cfg(feature = "build")]
-#[cfg_attr(docsrs, doc(cfg(feature = "build")))]
-#[expect(missing_docs, reason = "TODO")]
 pub mod built;
 
 #[cfg(feature = "build")]
@@ -30,3 +25,14 @@ pub mod deploy_provider;
 
 #[expect(missing_docs, reason = "TODO")]
 pub mod builder;
+
+#[cfg(stageleft_runtime)]
+#[cfg(feature = "trybuild")]
+#[cfg_attr(docsrs, doc(cfg(feature = "trybuild")))]
+#[expect(missing_docs, reason = "TODO")]
+pub mod trybuild;
+
+#[cfg(stageleft_runtime)]
+#[cfg(feature = "trybuild")]
+#[cfg_attr(docsrs, doc(cfg(feature = "trybuild")))]
+pub use trybuild::generate::init_test;
