--- hydro_lang/src/compile/rewrites/mod.rs
@@ -1,1 +1,0 @@
-pub mod properties;
