+++ hydro_lang/src/compile/trybuild/rewriters.rs
