+++ hydro_lang/src/live_collections/keyed_singleton.rs
@@ -1365,6 +1365,7 @@ where
     }
 }
 
+#[cfg(feature = "deploy")]
 #[cfg(test)]
 mod tests {
     use std::collections::HashMap;
