+++ hydro_lang/src/live_collections/keyed_stream/mod.rs
@@ -2059,6 +2059,7 @@ where
     }
 }
 
+#[cfg(feature = "deploy")]
 #[cfg(test)]
 mod tests {
     use futures::{SinkExt, StreamExt};
