+++ hydro_lang/src/live_collections/stream/mod.rs
@@ -2831,26 +2831,36 @@ where
 
 #[cfg(test)]
 mod tests {
+    #[cfg(feature = "deploy")]
     use futures::{SinkExt, StreamExt};
+    #[cfg(feature = "deploy")]
     use hydro_deploy::Deployment;
+    #[cfg(feature = "deploy")]
     use serde::{Deserialize, Serialize};
+    #[cfg(feature = "deploy")]
     use stageleft::q;
 
     use crate::compile::builder::FlowBuilder;
-    use crate::live_collections::stream::{ExactlyOnce, NoOrder, TotalOrder};
+    #[cfg(feature = "deploy")]
+    use crate::live_collections::stream::ExactlyOnce;
+    use crate::live_collections::stream::{NoOrder, TotalOrder};
     use crate::location::Location;
     use crate::nondet::nondet;
 
     mod backtrace_chained_ops;
 
+    #[cfg(feature = "deploy")]
     struct P1 {}
+    #[cfg(feature = "deploy")]
     struct P2 {}
 
+    #[cfg(feature = "deploy")]
     #[derive(Serialize, Deserialize, Debug)]
     struct SendOverNetwork {
         n: u32,
     }
 
+    #[cfg(feature = "deploy")]
     #[tokio::test]
     async fn first_ten_distributed() {
         let mut deployment = Deployment::new();
@@ -2883,6 +2893,7 @@ mod tests {
         }
     }
 
+    #[cfg(feature = "deploy")]
     #[tokio::test]
     async fn first_cardinality() {
         let mut deployment = Deployment::new();
@@ -2916,6 +2927,7 @@ mod tests {
         assert_eq!(external_out.next().await.unwrap(), 1);
     }
 
+    #[cfg(feature = "deploy")]
     #[tokio::test]
     async fn unbounded_reduce_remembers_state() {
         let mut deployment = Deployment::new();
@@ -2949,6 +2961,7 @@ mod tests {
         assert_eq!(external_out.next().await.unwrap(), 3);
     }
 
+    #[cfg(feature = "deploy")]
     #[tokio::test]
     async fn atomic_fold_replays_each_tick() {
         let mut deployment = Deployment::new();
@@ -2991,6 +3004,7 @@ mod tests {
         assert_eq!(external_out.next().await.unwrap(), (2, 6));
     }
 
+    #[cfg(feature = "deploy")]
     #[tokio::test]
     async fn unbounded_scan_remembers_state() {
         let mut deployment = Deployment::new();
@@ -3029,6 +3043,7 @@ mod tests {
         assert_eq!(external_out.next().await.unwrap(), 3);
     }
 
+    #[cfg(feature = "deploy")]
     #[tokio::test]
     async fn unbounded_enumerate_remembers_state() {
         let mut deployment = Deployment::new();
@@ -3059,6 +3074,7 @@ mod tests {
         assert_eq!(external_out.next().await.unwrap(), (1, 2));
     }
 
+    #[cfg(feature = "deploy")]
     #[tokio::test]
     async fn unbounded_unique_remembers_state() {
         let mut deployment = Deployment::new();
