+++ hydro_test/src/lib.rs
@@ -15,6 +15,6 @@ mod docs {
 mod test_init {
     #[ctor::ctor]
     fn init() {
-        hydro_lang::deploy::init_test();
+        hydro_lang::compile::init_test();
     }
 }
